+++ apps/backend/index.ts
@@ -1,9 +1,24 @@
 import fastify from 'fastify'
+import cors from '@fastify/cors'
 
 const server = fastify()
 
+server.register(cors, {
+  origin: '*',
+  methods: ['GET']
+})
+
 server.get('/ping', async (request, reply) => {
   return 'pong\n'
+})
+
+server.get('/people', async (request, reply) => {
+  return [
+    { name: 'Alice', age: 25, email: 'alice@example.com' },
+    { name: 'Bob', age: 30, email: 'bob@example.com' },
+    { name: 'Charlie', age: 35, email: 'charlie@example.com' },
+    { name: 'Jan', age: 24, email: 'johnnyciepiela@gmail.com' }
+  ]
 })
 
 server.listen({ port: 8080 }, (err, address) => {