+++ Changelog.md
@@ -1,3 +1,6 @@
 ## Task 1
 ### Faza Red
 - Added unit tests, which test fetching data about people, added tests for /people endpoint
+
+### Faza Green
+- Implemented functionality for fetching people data and displaying it in TaskOneView.vue component + fixed CORS issue